moved getRoleScopeId test
